создал статик мапперы в ДТО

+++ src/main/java/ru/skypro/homework/service/CommentServiceImpl.java
@@ -6,10 +6,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 import ru.skypro.homework.dto.CommentDTO;
 import ru.skypro.homework.dto.CommentInfoDTO;
+import ru.skypro.homework.dto.Role;
 import ru.skypro.homework.pojo.Ad;
 import ru.skypro.homework.pojo.Comment;
+import ru.skypro.homework.pojo.User;
 import ru.skypro.homework.repository.AdRepository;
 import ru.skypro.homework.repository.CommentRepository;
+import ru.skypro.homework.repository.UserRepository;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -20,6 +23,7 @@ public class CommentServiceImpl implements CommentService {
 
     private final CommentRepository commentRepository;
     private final AdRepository adRepository;
+    private final UserRepository userRepository;
 
     @Override
     public CommentDTO addCommentToAd(Long pk, CommentDTO commentDTO, Long userId) {
@@ -45,6 +49,10 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public List<CommentInfoDTO> getAllCommentsByPK(Long pk) {
+        if(adRepository.findByPk(pk) == null){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
         List<Comment> comments = commentRepository.findByPk(pk); // Получаем комментарии из базы данных
         List<CommentInfoDTO> commentsFullInfo = new ArrayList<>();
         for (Comment comment : comments) {
@@ -55,28 +63,37 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public String deleteCommentByAdAndCommentId(Long pk, Long commentId) {
+    public String deleteCommentByAdAndCommentId(String userName, Long pk, Long commentId) {
         Comment comment = commentRepository.findByPkAndCommentId(pk, commentId);
-
-        if (comment != null) {
-            commentRepository.delete(comment); // Удаляем комментарий из базы данных
-            return "Комментарий успешно удален";
-        } else {
-            return "Комментарий не найден";
+        User user = userRepository.findUserByUserName(userName);
+        if(user == null){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }else if(user.getUserID().equals(comment.getUserId()) || user.getRole().equals(Role.ADMIN)) {
+            if (comment != null) {
+                commentRepository.delete(comment); // Удаляем комментарий из базы данных
+                return "Комментарий успешно удален";
+            } else {
+                return "Комментарий не найден";
+            }
         }
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN);
     }
 
 
     @Override
-    public CommentInfoDTO updateCommentAndGetInfo(Long pk, Long commentId, CommentDTO commentDTO) {
+    public CommentInfoDTO updateCommentAndGetInfo(String userName, Long pk, Long commentId, CommentDTO commentDTO) {
         Comment comment = commentRepository.findByPkAndCommentId(pk, commentId);
-
-        if (comment != null) {
-            // Обновляем текст комментария
-            comment.setText(commentDTO.getText());
-            // Сохраняем обновленный комментарий в бд
-            comment = commentRepository.save(comment);
-            return CommentInfoDTO.fromComment(comment);
+        User user = userRepository.findUserByUserName(userName);
+        if(user == null){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }else if(user.getUserID().equals(comment.getUserId()) || user.getRole().equals(Role.ADMIN)) {
+            if (comment != null) {
+                // Обновляем текст комментария
+                comment.setText(commentDTO.getText());
+                // Сохраняем обновленный комментарий в бд
+                comment = commentRepository.save(comment);
+                return CommentInfoDTO.fromComment(comment);
+            }
         }
         throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
